requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
pygsl == 2.3.2
python_dotenv == 0.19.2
requests == 2.27.1
tweepy == 4.4.0
